Finished!
I finished up the level, adding Big to the end, and I edited the Hazardous.cs script so it also checks for the win condition.  All that's left to do is click "build."  My computer only lets me build for windows, so if someone else could do the Mac build, I'd appreciate it.